+++ app/controllers/admin_controller.rb
@@ -1,6 +1,4 @@
 class AdminController < ApplicationController
   before_action :authenticate_user!
   before_action :check_system_admin
-
-
 end