+++ app/models/user.rb
@@ -39,6 +39,7 @@ class User < ActiveRecord::Base
   has_many :comments, -> { where deleted: false }
   has_many :datasets, -> { where deleted: false }
   has_many :dataset_file_audits
+  has_many :hosting_requests, -> { current }
   has_many :reviews
   has_many :tags, -> { where deleted: false }
   has_many :tools