+++ test/models/hosting_request_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class HostingRequestTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end